requirements.txt
fastapi
uvicorn
azure-identity
azure-ai-projects
azure-ai-agents
